requirements.txt
click==8.0.0
Flask==2.0.0
itsdangerous==2.0.0
Jinja2==3.0.0
MarkupSafe==2.0.0
mysql-connector-python==8.0.25
protobuf==3.17.0
six==1.16.0
Werkzeug==2.0.0
